Add MIT License to the project
This commit adds the MIT License to the project, which allows others to freely use, modify, and distribute the code with minimal restrictions. The license also includes a disclaimer of warranty, making it clear that the software is provided "as-is".

The full text of the MIT License has been added to the `LICENSE` file in the root of the repository.